Documentation Fixes and Updates (#595)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Various fixes to the documentation.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Remove configuration for unused output formats for sphinx
- [x] Bump os version on ReadTheDocs
- [x] Bump Python version for documentation to 3.12 (GitHub Actions and
ReadTheDocs) -> I decided not to do this because I think the
documentation and deployment should be done with the minimum version of
Python that the library supports, as sort of a last line of defense to
check that we do not do anything that only works in new Python versions.
It should also be noted that bumping to 3.12 slightly changes the
documentation that shows up for AddStatus as it now shows all the
methods of `IntEnum`; I am unclear why this happens.
- [x] Add current year to copyright
- [x] Fix canonical links in layout.html
- [ ] Fix ReadTheDocs flyout not showing up -> Unsure how to do this;
switching to a new theme may fix this

### Upgrading Sphinx

I tried upgrading Sphinx since we are using a pretty old version, but it
seems that this introduces a host of problems because the
`sphinx-material` theme is no longer supported and thus does not support
the new Sphinx versions (see this issue from the maintainer here:
https://github.com/bashtage/sphinx-material/issues/141). There are also
a bunch of issues here:
https://github.com/bashtage/sphinx-material/issues -- many are related
to trying to use `sphinx-material` in new projects.

These were the subtasks that I tried:

- [x] I upgraded `sphinx-autobuild` at first but got errors similar to
what is seen here, so I kept it at a lower version:
https://github.com/sphinx-doc/sphinx-autobuild/issues/178
- [x] `autosummary` raised some complaints about not including the whole
module name so I updated the various module docstrings accordingly
- [x] See if this fixes the readthedocs flyout not showing up -> I tried
deploying with upgraded Sphinx and it did not help :(

It may also be possible to instead switch to the `sphinx-immaterial`
theme: https://github.com/jbms/sphinx-immaterial/

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `isort` and `ruff`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go