Migrate docs to sphinx-immaterial theme (#596)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Since sphinx-material is no longer supported, this PR migrates the
documentation to sphinx-immaterial
https://github.com/jbms/sphinx-immaterial

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Update Sphinx and other deps; install sphinx-immaterial theme
- [x] Check ReadTheDocs flyout works -> it does!
- [x] Copy button for code -> `content.code.copy` button
- [x] Check dark mode -> works!
- [x] Tweak TOC -- put Usage and Reference stuff at top level; expansion
is not nice because it's messy
- [x] How to pass things to the jinja context? Is there a `sphinx-jinja`
extension that we need? What is the `html_context`?
- [x] Tweak custom CSS rules and remove ones that are no longer needed
- [x] Work through config settings
- [x] How to allow the theme to be wider? ->
https://github.com/squidfunk/mkdocs-material/discussions/6404
- [x] Fix templates
  - [x] Special documentation for classes
  - [x] Canonical links and opengraph links
- [x] Open in Colab links -> wrote a small extension for this in
`docs/_ext`
- [x] Fix autosummary warnings -> removed module names in the
autosummary definitions in the docstrings
- [x] We get many warnings that look like `Parameter name XXX does not
match any of the parameters defined in the signature`. This is due to
using classes as default values, which is hard to parse (see here:
https://github.com/jbms/sphinx-immaterial/issues/140). This cannot
really be worked around so it will just be left as is.
- [x] We are getting warnings from sphinx-codeautolink related to the
use of linenos in the examples; it seems to be similar to what is shown
here but I am not able to make a minimal reproducible example of it:
https://github.com/felix-hilden/sphinx-codeautolink/issues/137
- [x] Remove sphinx-codeautolink since it seems to be introducing many
warnings in the build, and it is anyway not super obvious that you
should click on the links in the documentation. Some of the links are
also wrong, like `ribs.archives._grid_archive.GridArchive` instead of
`ribs.archives.GridArchive`
- [x] Minify HTML -- used the `minify-html` package and wrote an
extension in `docs/_ext`; had issues with failures on v0.16.4 so used
v0.15.0; see here: https://github.com/wilsonzlin/minify-html/issues/225

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `isort` and `ruff`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go