Proof of concept for building Latex docs (#601)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

I was experimenting with building the docs as a Latex file, since Sphinx
provides the ability to build many different documentation formats. Most
of the other formats like man pages, epub, dirhtml, singlehtml all work
out of the box (although they need some tweaking to get the right
appearance), while latex errors out due to a variety of issues. The TODO
section below lists the steps I took to get it to mostly work. I think
if we wanted this to work properly, it would be possible with a lot more
polishing. I have decided to merge this PR since the steps here all seem
to not affect the current code, with the exception of item 6 (commenting
out the gif; I reverted this).

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

1. **Make all tables have a table body.** It appears that there is a
sphinx transform that will throw an error like:

    ```
    Exception occurred:
File
"/usr/local/lib/python3.11/site-packages/sphinx/builders/latex/transforms.py",
line 441, in depart_table
        tbody = next(node.findall(nodes.tbody))
                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    StopIteration
    ```
This issue is shown here:
https://github.com/sphinx-doc/sphinx/issues/11751. The solution listed
in that issue appears to be to make all tables have a table body; as
such, I added a table body to the tables in `docs/whats-new/v0.5.0.md`.
I debugged this by modifying the `transforms.py` file to print out the
node on line 441, so that I could try to determine where the table with
the missing tbody was located.

2. **Latex does not support SVG images with `\includegraphics`.** As
such, I had to install the `svg2pdfconverter` extension shown here:
https://pypi.org/project/sphinxcontrib-svg2pdfconverter/ This extension
converts the SVG files to PDFs with one of many possible commands; I
chose to use Inkscape. Thus, the extension is added in `docs/conf.py` as
`sphinxcontrib.inkscapeconverter`. This does not appear to affect the
output when using the regular `html` output, i.e., the PDFs are _only_
generated for latex. Relevant StackOverflow question
[here](https://stackoverflow.com/questions/49898798/using-sphinx-extension-to-convert-svg-to-pdf).
Info on how it works is available in the source code
[here](https://github.com/missinglinkelectronics/sphinxcontrib-svg2pdfconverter/blob/master/sphinxcontrib/inkscapeconverter.py)
and in the reference image converter extension
[here](https://www.sphinx-doc.org/en/master/usage/extensions/imgconverter.html),
as well as the base class for image converters
[here](https://www.sphinx-doc.org/en/master/_modules/sphinx/transforms/post_transforms/images.html).

3. **Bad file paths for images generated by svg2pdfconverter.** When
`svg2pdfconverter` runs, one of the files it generates is called
`pyribs.pdf` since one of the images in the README is called
`pyribs.svg`. Thus, I created a custom build command (see `Makefile`)
that builds the pdf from the compiled latex in `docs/_build/pdf/`.

4. **pdflatex does not support Unicode characters.** Due to showing
progress bars from tqdm, the tutorials include Unicode characters, which
are not supported in pdflatex. Thus, I set `latex_engine = "xelatex"` in
`docs/conf.py`, since xelatex handles Unicode.

5. **List too deeply nested.** It turns out there is a docstring list in
`ArrayStore` that is too deeply nested, so I modified that docstring to
have a less deeply nested list.

6. **(UNRESOLVED) Latex raises an error due to the presence of a gif.**
I commented out this gif in `ns_maze.ipynb`, but we would need some
extension to convert this gif to make a Latex build work, or we would
need a way to ignore the image for Latex.

7. **(UNRESOLVED) Various errors and broken references.** When I left
off, there were still a ton of errors in the Latex build. While the PDF
built, it seems many issues still need to be resolved.

The final instructions are as follows:

```
make latexdocs
```

In debugging this, it was helpful to run:

```
cd docs
make latex
cd _build
make
```

This way, I could repeatedly go through the Latex errors myself instead
of having Sphinx regenerate the whole docs. I encountered errors like
the ones shown
[here](https://tex.stackexchange.com/questions/331296/error-pdf-file-is-damaged-attempting-to-reconstruct-xref-table),
[here](https://tex.stackexchange.com/questions/507213/package-incompatibility-with-hyperxmp-and-silence),
and
[here](https://tex.stackexchange.com/questions/669494/latexmk-fls-file-lists-log-file-as-an-input-file-and-using-biber-omitting-te).

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `isort` and `ruff`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [N/A] I have added a one-line description of my change to the
changelog in
      `HISTORY.md`
- [x] This PR is ready to go