Replace isort with ruff import check (#603)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Previously, we used isort to sort the imports and ruff to handle the
rest of the formatting. This PR migrates to using ruff for the import
sorting as well, which reduces the number of tools we depend on.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Update pyproject.toml -- the settings for ruff's import sorting
are similar to isort and are placed under `tool.ruff.lint.isort`
- [x] Update CONTRIBUTING steps and PR template
- [x] Format codebase with the new import sorting -- imports in notebook
tutorials were modified

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `ruff`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go