Fix existing type annotation errors in the codebase (#620)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Before adding type annotations in #611, this PR fixes some of the
existing errors that show up when running `ty check`

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Mark arguments as `Collection[str]` rather than `list[str]` in
QDAIF tutorial
- [x] Ignore display_in_notebook not being found in QDAIF tutorial
- [x] Ignore error in `benchmarks/cvt_add.py`
- [x] Fix `upscale_schedule` type in `examples/bop_elites.py` -- see
here for how tuples should be annotated; it's not the same as lists:
https://docs.python.org/3/library/typing.html#annotating-tuples
- [x] Expand types for `CQDScoreResult`
- [x] Ignore unresolved import for cKDTree in CVTArchive and
ProximityArchive
- [x] Ignore `call-non-callable` when calling `aggregate` -- this
happens because `ty` sees that `aggregate` might be None, which is not
callable. `aggregate` might be None because `numpy-groupies` sets
`aggregate_nb` based on whether `numba` is installed -- if it is not
installed, `aggregate_nb` is None. Related issues:
https://github.com/astral-sh/ty/issues/309,
https://github.com/astral-sh/ty/issues/708
- A followup PR will make a more elegant fix where we check if
`aggregate_nb` can be imported and default to a regular version of
`aggregate`
- [x] Ignore unresolved-import for cupy_backends in `tests/conftest.py`
- [x] Force type for new_data in SlidingBoundariesArchive.remap

Remaining errors come from the Lunar Lander example and tutorial (due to
use of Gymnasium), the QDHF tutorial (due to archive potentially being
None), or ArrayStore (due to dynamic return types in `retrieve`.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [x] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go