Raise `KeyError` in `ArchiveDataFrame.get_field` when field not found (#626)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Previously, when a field was not found in an ArchiveDataFrame, we would
just return `None`. However, having a None value is not very useful and
can lead to unexpected behavior. This PR instead raises a KeyError
similar to what one would get if accessing an invalid column in a
regular DataFrame.

This PR also adds some type annotations to ArchiveDataFrame.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Raise KeyError in ArchiveDataFrame
- [x] Add test to check KeyError (replaced old test for checking if
value is None)

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [x] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go