Annotate ArrayStore and add types for Array API (#629)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Annotations for the ArrayStore. Since we support the Array API in the
ArrayStore, this PR also creates types to make it easier to work with
the Array API.

Regarding typing of `xp`, official types for Array API are coming soon:
https://github.com/data-apis/array-api-typing It also seems there is an
unofficial package here that gets the job done:
https://github.com/34j/types-array-api We should install the official
package once it is available and update the files in this PR
accordingly. I added a note to the tracking issue #624.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Add annotations to ArrayStore
- [x] Add new types in `ribs.typing`: `Array`, `DType`, `Device`,
`ArrayVar`
- [x] Modify documentation to require importing torch and cupy so that
the above type annotations will work
- [x] Check if it is possible to avoid those imports, perhaps by using
`if TYPE_CHECKING` -> Yes, is it possible by doing something like below.
However, now Sphinx doesn't see the types correctly in the
documentation, so it's easier to just add an extra that installs these
dependencies. I added the `dev-array-api` extra.
```
if TYPE_CHECKING:
    import torch
    import cupy as cp

Array = np.ndarray
if TYPE_CHECKING:
    Array = Union[np.ndarray, torch.Tensor, cp.ndarray]
```
- [x] Add types to `arr_readonly` in utils

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [x] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go