Annotate archives (#634)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Annotate remaining archives. We mostly assume things are still written
in numpy, and will update later as Array API is added.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] CategoricalArchive
- [x] CVTArchive
- [x] DensityArchive
- [x] GridArchive
- [x] ProximityArchive
- [x] SlidingBoundariesArchive
- [x] Copy solution_dim from ArchiveBase to other archives
- [x] Fix headers in documentation; they don't show up very nicely -- it
seems this is caused by having `np.float64` as the default value. So I
added an autodoc extension in `conf.py` that replaces the `np.float64`
repr() with simple text. Now the docstrings show up fine. The final
result isn't as good as it could be because `DTypeLike` gets expanded,
but it's much better than before.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [x] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go