Allow pickling classes that have the `_xp` array module attribute (#644)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Previously, `ArrayStore` could not be pickled because it had `self._xp`
which was an array module like torch or numpy, and modules cannot be
pickled. This PR fixes that issue by adding a mixin that modifies the
`__getstate__` and `__setstate__` methods to make pickle play nicely
with `self._xp`.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Add test for ArrayStore pickling
- [x] Implement `PickleXPMixin` and add to ArrayStore

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [x] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go