Remove `centroid_method` from `CVTArchive` and add centroid generation tutorial (#662)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

The `centroid_method` parameter in `CVTArchive` takes up a lot of space
in the `__init__`, and I do not find it to be particularly flexible. As
such, I think it is more useful to instead demonstrate different ways of
generating centroids and have that as a reference for users. Hence, this
PR adds a new small tutorial on how to do this.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Remove `centroid_method` from `CVTArchive`
- [x] Remove `centroid_method` in tests
- [x] Add `centroid_methods.ipynb` tutorial

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [x] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go