Fix CI errors (#667)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Have been getting errors with Windows and the examples lately.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [FAILED] Fix error on windows-latest that says: "This probably means
that tk wasn't installed properly."
  - Related: https://github.com/actions/setup-python/issues/1102
- This error appears to happen only on Windows, and only for the
visualize tests. Furthermore, it happens on the cvt_archive_3d_plot
tests, but not on the 2D visualizations.
  - I am unsure how to fix this; it may be best to just comment it out?
- [x] Fix SWIG issues in examples by using system SWIG from GitHub
Actions
- [x] Comment out the "Remove swig" step in examples and tutorials ->
Previously, we removed SWIG so that we could test the installation
process, but this has turned out to be quite fussy.
- [x] Modify `examples/lunar_lander.py` so that swig is no longer
installed from PyPI. This way, we use the system's SWIG rather than
trying to install SWIG on our own. It appears the GitHub Actions
container already has SWIG installed, so we can just use that.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [N/A] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go