Separate centroid generation from init in CVTArchive (#663)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

This PR introduces various breaking changes that fully separate the
creation of centroids from the CVTArchive. The motivation is that
CVTArchive currently takes in tons of arguments just to generate the
centroids, but this makes the init method very crowded. The focus of
CVTArchive should be that it takes in centroids and forms an archive
from the centroids.

Previously, #665 added a function that can create centroids using
k-means clustering (`k_means_centroids`). Among other things, this PR
updates CVTArchive to now call this function. Various other parameters
for CVTArchive are also deprecated to simplify the API.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] **DEPRECATE** `cells` and `custom_centroids` in CVTArchive; these
are now replaced with a new `centroids` parameter
- [x] **DEPRECATE** the samples property in CVTArchive -- since
`k_means_centroids` now generates samples, there is no need to store it
in CVTArchive
- [x] **DEPRECATE** `plot_samples` in `cvt_archive_heatmap` and
`cvt_archive_3d_plot` -- since CVTArchive no longer has samples, they
cannot be plotted in the heatmap function
- [x] Fix tests
  - [x] Revise CVTArchive tests (in `tests/archives`)
    - [x] Test shape for `centroids` parameter
    - [x] Check regular centroids -- shape, dtypes, etc.
  - [x] Update other breaking tests and tutorials
  - [x] Fix benchmark code
- [x] Update centroid methods tutorial with example usage

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [x] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go