Allow visualize tests to fail on windows-latest (#671)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

We have been getting a ton of windows errors with "This probably means
that tk wasn't installed properly."; see #667 for more info. It used to
be pretty inconsistent, but now it seems to be a persistent error. This
PR allows to continue-on-error for this specific step that fails. The
specific step is the testing of ribs.visualize, where running
`cvt_archive_3d_plot` seems to fail with the aforementioned error.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [N/A] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go