Replace `ckdtree_kwargs` with `kdtree_kwargs` in CVTArchive and ProximityArchive (#676)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

In #669, we previously chose to keep the name `ckdtree_kwargs`. However,
I think not a lot of people use `ckdtree_kwargs`, so it is alright to
introduce this breaking change to rename it to `kdtree_kwargs`, which
makes the overall usage more consistent since scipy has deprecated
`cKDTree`.

As an illustration of the rarity of this argument, we actually do not
use `ckdtree_kwargs` in our tests.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Deprecate `ckdtree_kwargs` in CVTArchive
- [x] Deprecate `ckdtree_kwargs` in ProximityArchive

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [x] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go