Allow specifying `centroids` with filenames in CVTArchive (#679)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

This PR makes it possible to pass in file paths for the `centroids`
parameter in `CVTArchive`. When the path is passed in, CVTArchive reads
from it with `np.load` and uses the contents of the file as the
centroids.

Note that `np.load` does allow using byte streams like file objects as
the file, but I decided not to include that in the options for
centroids, as it seems complicated to distinguish between byte streams
and array-like objects (but I could be wrong).

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Implement file path for `centroids` -- this required rearranging
the init method in CVTArchive since the ArrayStore depends on having the
centroids available for determining the capacity. Thus, the centroids
must first be read before the ArrayStore can be created.
- [x] Add test

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [x] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go